Fix travis build
Travis is reporting ENOENT errors because of missing parent
directories for the sync path. It is working locally, though,
and it should work since the gem is creating them using mkdir_p.

This is a dirty fix, it ensures that error will not occur, but this should not
be necessary.