Revert "Added encounter id and encounter uuid to core fields"
CDP is broken (as of 13986a577f0f8fbdd6aa8f93fb9e7c9548fa7584)
because it still doesn't handle encounter_id and this fields,
but I need to update it's dependency to cdx_core for implementing
patient_age as a duration (see instedd/cdp#245).

The reverted commit should be re-applied when cdp learns how to
handle it.

This reverts commit 57c2d688bed69640c32c55cfef4f64f31e554928.