Added the controversial db per branch setting
If `db_per_branch` setting is truthy, then the development database will be named `cdp_development_BRANCH`, where BRANCH is the current branch name. When switching to a new branch, running rake db:clone will copy the current cdp_development database to a new cdp_development_newbranch database. This allows to easily switch between different branches, without causing issues with migrations specific to a branch having been run on the development database, but that are not expected in another branch.