Support attachments in non-default S3 regions
Uploading attachments crashed with an exception due to an issue with S3.

```
AWS::S3::Errors::PermanentRedirect (The bucket you are attempting to access must be addressed using the specified endpoint. Please send all future requests to this endpoint.)
```

We now allow to specify the region-specific endpoint of the S3 bucket to avoid this.