Disable whenever init in Docker image build
It currently needs the whole Rails app to be up, which
is not true while building the Docker image.

This commit disables the initialization of the crontab during
builds, so `main` builds once again, and we'll keep working to
add the initialization soon - in a way that works.

See #1351 (done for #1305 and #1331)
See 827db4651272ad2f46fd57889eb812b9ffcf43cb