Spec: use associations directly in factories
There was caused where we passed an `user_id: user.id` instead of
being direct by passing the associated record to set directly (for
example `user: user`).

Passing the records directly seem to speed the spec suite runs,
which may be caused by Machinist not creating the association just
to override it's id afterwards (thus not using it).