Spec: introduce setup_fixtures helper to share records
Extracts what was previously done on the Blender specs, where we
create records once before :all, and we memoize them as instance
variables that were reloaded before :each spec.

This patch follows the same logic, but uses a generic
`#setup_fixtures` method that takes a block that is must create
records and save them as ivars). The method will eventually create
as many `#let!` definitions as the block created ivars.