Add code coverage
This is to understand where and what tests are missing in the
application, to have a global understanding of the test quality of
the application.

As of this commit the overall coverage is 80%, which looks great,
but looking more closely we see that 9 controllers are 0% hence
totally untested, and 6 controllers are below 70% coverage.