Migrate Capybara to use Selenium WebDriver
PhantomJS is really too old. It's even older than IE11. It doesn't
support ES5, and we must use es5-shim just for tests to pass.

We must move on to something a little more recent. This patch
changes the driver to Selenium WebDriver and installs Firefox into
the Ruby development image. It's still an old Firefox 68 but it's
still more recent.