Remote Selenium service (distinct compose service)
It avoids bloating the web service with unrelated software (namely
firefox and geckodriver), it also allows us to have a recent and
modern browsers, for example Firefox 91 ESR (Extended Support).

Note that we must mount the source folder so we can attach and
upload fixture files in the remote browser during tests!