Fix: copy QC info for already transferred sample (#1581)
* Fix: use scope instead of building query in controller
* Fix: allow copy qc info for transferred sample

A sample may have been already transferred and have QC info attached
to it. That same sample may be transferred again, and we musn't say
that the sample doesn't have any QC info to attach nor shall we
disable the ability to copy the info.

Note that this actually does nothing: the QC info will be kept on
the sample, even if the user doesn't check the copy QC info.