1598 show date last modified date in samples (#1739)
* Added Modified column with sort and filter

* coloring ordered

* dropdown Modified filter

* Update app/controllers/samples_controller.rb

Co-authored-by: Julien Portalier <julien@portalier.com>

* fixed inline condition and put back the column name I accidentally changed

* input date placeholder color working

* filled input date fields with filled style on page ready

* controlling columns that can be sorted

* Fix Sort Filter Present

* Fix Sort Filter Present

Checking if the parameter is present before
checking if it's inside the sort_column list

* Fix CSS Input Date For All Browsers

I was changing the color for Chrome only. So I removed
the webkit selector. I put a comment inside input date
javascript too, to describe the problem and prevent
other developers to waste time with this.

Co-authored-by: Carlos Zillner <czillner@manas.com.ar>
Co-authored-by: Julien Portalier <julien@portalier.com>