Rework BatchesSelector component as BoxBatchesForm
The component isn't generic but tied to the boxes#create action, and
goes much further than selecting a list of batches: for each batch we
select.

Extracts BoxBatchForm to handle the individual form for each form, with
one to many concentrations.

The React component now merely generates a plain form, never really
removing the actual inputs from the DOM, so they're always sent with the
form (no need for duplicated hidden inputs).

Last but not least, allows to reopen a batch form.