GHA: cancel previous action runs
This should avoid race conditions when merging multiple pull requests in
a row, where the action from a previous run would finish after the
action from the last merge, and thus push a docker image that isn't the
latest (but will still act as the latest.

It would also avoid building & pushing multiple docker images when only
the last one is interesting and ought to be deployed.

We also don't need to keep running previous tests in a branch when we
push a new commit.